pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "baal"
version = "0.1.0"
description = "Telegram bot platform for deploying AI agents on Aleph Cloud with LibertAI inference"
requires-python = ">=3.12"
dependencies = [
    "pydantic-settings>=2.0",
    "aiosqlite>=0.20",
    "cryptography>=42.0",
]

[project.optional-dependencies]
core = [
    "httpx>=0.27",
    "aleph-sdk-python[ethereum]",
    "eth_account",
    "aiosqlite>=0.20",
]
bot = [
    "baal[core]",
    "python-telegram-bot[ext]>=22.0",
]
agent = [
    "fastapi>=0.109",
    "uvicorn>=0.27",
    "openai>=1.0",
    "httpx>=0.27",
    "aiosqlite>=0.20",
]
api = [
    "baal[core]",
    "fastapi>=0.109",
    "uvicorn>=0.27",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.29",
    "alembic>=1.13",
    "python-jose[cryptography]>=3.3",
    "authlib>=1.3",
    "itsdangerous>=2.1",
    "web3>=6.0",
    "pydantic[email]>=2.0",
    "resend>=0.7",
    "httpx>=0.27",
]
dev = [
    "ruff",
    "pytest",
    "pytest-asyncio",
]

[tool.hatch.build.targets.wheel]
packages = ["src/baal", "src/baal_agent", "src/baal_core", "src/liberclaw"]
